+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Company Certificate
   - Provide new link from user menu for the company certificate page
   - UI components created to display filter, sort and company certificates
+  - Company Certificate Details overlay
 - Company Roles
   - Fetch the standard library data from standards.json and implement Table component to display it in the company roles section.